feat: build js
