+++ assets/js/package.json
@@ -4,12 +4,26 @@
   "description": "contact form 7 huge file upload",
   "scripts": {
     "test": "yarn lint",
-    "lint": "eslint ./*.js"
+    "lint": "eslint ./*.js",
+    "build": "webpack --mode production --config webpack.config.js"
   },
   "author": "Technote <technote.space@gmail.com> (https://technote.space)",
   "license": "GPL-3.0",
   "devDependencies": {
+    "@babel/core": "^7.5.5",
+    "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
+    "@babel/plugin-transform-modules-commonjs": "^7.5.0",
+    "@babel/plugin-transform-react-jsx": "^7.3.0",
+    "@babel/plugin-transform-runtime": "^7.5.5",
+    "@babel/polyfill": "^7.4.4",
+    "@babel/preset-env": "^7.5.5",
+    "@babel/register": "^7.5.5",
+    "babel-loader": "^8.0.6",
+    "duplicate-package-checker-webpack-plugin": "^3.0.0",
     "eslint": "^6.1.0",
-    "eslint-plugin-react": "^7.14.3"
+    "eslint-plugin-react": "^7.14.3",
+    "speed-measure-webpack-plugin": "^1.3.1",
+    "webpack": "^4.36.1",
+    "webpack-cli": "^3.3.6"
   }
 }
